@@ -14,19 +14,23 @@
     "less-loader": "^2.2.0",
     "history": "2.0.0",
     "isomorphic-style-loader": "0.0.10",
-    "react": "0.14.7",
-    "react-dom": "0.14.7",
+    "react": "^15.0.1",
+    "react-dom": "^15.0.1",
     "react-routing": "0.0.7"
   },
   "devDependencies": {
     "assets-webpack-plugin": "^3.3.0",
-    "babel-cli": "^6.6.0",
+    "babel": "^6.5.2",
+    "babel-cli": "^6.9.0",
+    "babel-core": "^6.9.1",
     "babel-loader": "^6.2.4",
     "babel-plugin-react-transform": "^2.0.0",
     "babel-plugin-transform-runtime": "^6.6.0",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
+    "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-stage-0": "^6.5.0",
+    "babel-runtime": "^6.6.1",
     "core-js": "^2.1.3",
     "css-loader": "^0.23.1",
     "express": "^4.13.4",
@@ -35,6 +39,8 @@
     "json-loader": "^0.5.4",
     "less": "^2.5.1",
     "less-loader": "^2.2.0",
+    "react": "^15.0.1",
+    "react-dom": "^15.0.1",
     "react-transform-catch-errors": "^1.0.2",
     "react-transform-hmr": "^1.0.2",
     "url-loader": "^0.5.7",
